+++ src/main/java/shop/gaship/gashipshoppingmall/config/DataProtectionConfig.java
@@ -4,8 +4,6 @@ import java.util.Objects;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.reactive.function.client.WebClient;
 import shop.gaship.gashipshoppingmall.dataprotection.dto.SecureKeyResponse;
 import shop.gaship.gashipshoppingmall.dataprotection.exception.NotFoundDataProtectionReposeData;
@@ -37,14 +35,8 @@ public class DataProtectionConfig {
                 .blockOptional()
                 .orElseThrow(() -> new NotFoundDataProtectionReposeData(errorMessage))
                 .getBody())
-            .getBody()
-            .getSecret();
-    }
-
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
+                .getBody()
+                .getSecret();
     }
 
     public String getUrl() {
