+++ src/main/java/shop/gaship/gashipshoppingmall/member/advisor/MemberAdvisor.java
@@ -1,6 +1,5 @@
 package shop.gaship.gashipshoppingmall.member.advisor;
 
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -18,7 +17,6 @@ import shop.gaship.gashipshoppingmall.message.ErrorResponse;
  * @since 1.0
  */
 @RestControllerAdvice(basePackageClasses = MemberMarker.class)
-@Slf4j
 public class MemberAdvisor {
     /**
      * 회원과 관련하여 예상된 예외를 처리하는 메서드입니다.
@@ -26,12 +24,12 @@ public class MemberAdvisor {
      * @param exception 예외 객체입니다.
      * @return 예외의 메세지가 들어있는 객체를 반환합니다.
      */
-    @ExceptionHandler({SignUpDenyException.class, MemberNotFoundException.class})
+    @ExceptionHandler({ SignUpDenyException.class, MemberNotFoundException.class })
     public ResponseEntity<ErrorResponse> memberExceptionAdvice(RuntimeException exception) {
         return ResponseEntity
-            .badRequest()
-            .contentType(MediaType.APPLICATION_JSON)
-            .body(new ErrorResponse(exception.getMessage()));
+                .badRequest()
+                .contentType(MediaType.APPLICATION_JSON)
+                .body(new ErrorResponse(exception.getMessage()));
     }
 
     /**
@@ -44,10 +42,10 @@ public class MemberAdvisor {
      * @param exception 예외 객체입니다.
      * @return 예외의 메세지가 들어있는 객체를 반환합니다.
      */
-    @ExceptionHandler({Exception.class, RuntimeException.class})
+    @ExceptionHandler({ Exception.class, RuntimeException.class })
     public ResponseEntity<ErrorResponse> otherExceptionAdvice(Exception exception) {
         return ResponseEntity.internalServerError()
-            .contentType(MediaType.APPLICATION_JSON)
-            .body(new ErrorResponse(exception.getMessage()));
+                .contentType(MediaType.APPLICATION_JSON)
+                .body(new ErrorResponse(exception.getMessage()));
     }
 }
