+++ src/main/java/shop/gaship/gashipshoppingmall/member/dto/request/MemberCreationRequestOauth.java
@@ -1,4 +1,4 @@
-package shop.gaship.gashipshoppingmall.member.dto;
+package shop.gaship.gashipshoppingmall.member.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDate;
